requirements.txt
feedparser
requests
beautifulsoup4
openai
jinja2
python-dotenv
APScheduler